"N/A"

+++ python-server_side_rendering/task_00_intro.py
@@ -10,16 +10,17 @@ def generate_invitations(template, attendees):
         return
     for index, element in enumerate(attendees):
         file = "output_{}.txt".format(index + 1)
-        with open(file, 'w+') as f:
-            f.write(template.format(name = checker("name", element),
-                                    event_title = checker("event_title", element),
-                                    event_date = checker("event_date", element),
-                                    event_location = checker("event_location", element)))
+        if os.path.exists(file) is False:
+            with open(file, 'w+') as f:
+                f.write(template.format(name = checker("name", element),
+                                        event_title = checker("event_title", element),
+                                        event_date = checker("event_date", element),
+                                        event_location = checker("event_location", element)))
 
 def checker(key, dict):
     if key not in dict:
         return "{}: N/A".format(key)
     elif not dict[key]:
-        return "N/A"
+        return "\"N/A\""
     else:
         return "{}".format(dict[key])
